Only write config if not yet present

+++ write_config.py
@@ -126,27 +126,36 @@ def write_initial_config(search_path=None):
     if search_path is not None:
         search_path = os.path.expanduser(search_path)
 
-    print("Write config with search path {}".format(search_path))
-    config = configparser.ConfigParser()
-    get_initial_config(config, search_path)
     path = os.path.join(thisfilepath, "prophesy", "prophesy.cfg")
-    logging.info("Writing config to " + path)
-    with open(path, 'w') as configfile:
-        config.write(configfile)
+    if os.path.isfile(path):
+        print("Config file {} already exists.".format(path))
+    else:
+        print("Write config with search path {}".format(search_path))
+        config = configparser.ConfigParser()
+        get_initial_config(config, search_path)
+        logging.info("Writing config to " + path)
+        with open(path, 'w') as configfile:
+            config.write(configfile)
 
-    config = configparser.ConfigParser()
-    get_initial_dependencies_config(config)
     path = os.path.join(thisfilepath, "prophesy", "dependencies.cfg")
-    logging.info("Writing dependencies to " + path)
-    with open(path, 'w') as configfile:
-        config.write(configfile)
+    if os.path.isfile(path):
+        print("Dependencies file {} already exists.".format(path))
+    else:
+        config = configparser.ConfigParser()
+        get_initial_dependencies_config(config)
+        logging.info("Writing dependencies to " + path)
+        with open(path, 'w') as configfile:
+            config.write(configfile)
 
-    config = configparser.ConfigParser()
-    get_initial_web_config(config)
     path = os.path.join(thisfilepath, "prophesy_web", "prophesy_web.cfg")
-    logging.info("Writing config to " + path)
-    with open(path, 'w') as configfile:
-        config.write(configfile)
+    if os.path.isfile(path):
+        print("Web config file {} already exists.".format(path))
+    else:
+        config = configparser.ConfigParser()
+        get_initial_web_config(config)
+        logging.info("Writing config to " + path)
+        with open(path, 'w') as configfile:
+            config.write(configfile)
 
 
 if __name__ == "__main__":
